Move nodes API requests into the `.Specification.NodesApi` namespace
