+++ src/OpenSearch.Client/Cluster/ClusterStats/ClusterStatsResponse.cs
@@ -27,6 +27,7 @@
 */
 
 using System.Runtime.Serialization;
+using OpenSearch.Client.Specification.NodesApi;
 
 namespace OpenSearch.Client.Specification.ClusterApi
 {
