+++ src/OpenSearch.Client/Nodes/NodesInfo/NodeInfo.cs
@@ -31,7 +31,7 @@ using System.Runtime.Serialization;
 using OpenSearch.Net;
 using OpenSearch.Net.Utf8Json;
 
-namespace OpenSearch.Client
+namespace OpenSearch.Client.Specification.NodesApi
 {
 	[DataContract]
 	public class NodeInfo
