+++ src/OpenSearch.Client/Nodes/NodesResponseBase.cs
@@ -26,12 +26,31 @@
 *  under the License.
 */
 
-namespace OpenSearch.Client
+using System.Collections.Generic;
+using System.Runtime.Serialization;
+using OpenSearch.Net;
+
+namespace OpenSearch.Client.Specification.NodesApi
 {
-	[MapsApi("nodes.info.json")]
-	public partial interface INodesInfoRequest { }
+	public abstract class NodesResponseBase : ResponseBase
+	{
+		[DataMember(Name = "_nodes")]
+		public NodeStatistics NodeStatistics { get; internal set; }
+	}
 
-	public partial class NodesInfoRequest { }
+	[DataContract]
+	public class NodeStatistics
+	{
+		[DataMember(Name = "failed")]
+		public int Failed { get; internal set; }
 
-	public partial class NodesInfoDescriptor { }
+		[DataMember(Name = "successful")]
+		public int Successful { get; internal set; }
+
+		[DataMember(Name = "total")]
+		public int Total { get; internal set; }
+
+		[DataMember(Name = "failures")]
+		public IReadOnlyCollection<ErrorCause> Failures { get; internal set; } = EmptyReadOnly<ErrorCause>.Collection;
+	}
 }
