+++ src/OpenSearch.Client/Nodes/NodesStats/AdaptiveSelectionStats.cs
@@ -28,7 +28,7 @@
 
 using System.Runtime.Serialization;
 
-namespace OpenSearch.Client
+namespace OpenSearch.Client.Specification.NodesApi
 {
 	[DataContract]
 	public class AdaptiveSelectionStats
