+++ src/OpenSearch.Client/Nodes/NodesStats/Statistics/NodeIngestStats.cs
@@ -26,12 +26,23 @@
 *  under the License.
 */
 
-namespace OpenSearch.Client
+using System.Collections.Generic;
+using System.Runtime.Serialization;
+using OpenSearch.Net;
+using OpenSearch.Net.Utf8Json;
+
+namespace OpenSearch.Client.Specification.NodesApi
 {
-	[MapsApi("nodes.reload_secure_settings.json")]
-	public partial interface IReloadSecureSettingsRequest { }
+	public class NodeIngestStats
+	{
+		/// <summary> Per pipeline ingest statistics </summary>
+		[DataMember(Name = "pipelines")]
+		[JsonFormatter(typeof(VerbatimInterfaceReadOnlyDictionaryKeysFormatter<string, IngestStats>))]
+		public IReadOnlyDictionary<string, IngestStats> Pipelines { get; internal set; }
+			= EmptyReadOnly<string, IngestStats>.Dictionary;
 
-	public partial class ReloadSecureSettingsRequest { }
-
-	public partial class ReloadSecureSettingsDescriptor { }
+		/// <summary> Overall global ingest statistics </summary>
+		[DataMember(Name = "total")]
+		public IngestStats Total { get; set; }
+	}
 }
