+++ src/OpenSearch.Client/Nodes/ReloadSecureSettings/ReloadSecureSettingsResponse.cs
@@ -26,12 +26,20 @@
 *  under the License.
 */
 
-namespace OpenSearch.Client
+using System.Collections.Generic;
+using System.Runtime.Serialization;
+using OpenSearch.Net;
+using OpenSearch.Net.Utf8Json;
+
+namespace OpenSearch.Client.Specification.NodesApi
 {
-	[MapsApi("nodes.stats.json")]
-	public partial interface INodesStatsRequest { }
+	public class ReloadSecureSettingsResponse : NodesResponseBase
+	{
+		[DataMember(Name = "cluster_name")]
+		public string ClusterName { get; internal set; }
 
-	public partial class NodesStatsRequest { }
-
-	public partial class NodesStatsDescriptor { }
+		[DataMember(Name = "nodes")]
+		[JsonFormatter(typeof(VerbatimInterfaceReadOnlyDictionaryKeysFormatter<string, NodeStats>))]
+		public IReadOnlyDictionary<string, NodeStats> Nodes { get; internal set; } = EmptyReadOnly<string, NodeStats>.Dictionary;
+	}
 }
