+++ tests/Tests/ClientConcepts/Connection/ConnectionReuseAndBalancing.cs
@@ -38,6 +38,7 @@ using OpenSearch.OpenSearch.Xunit.XunitPlumbing;
 using OpenSearch.Net;
 using FluentAssertions;
 using OpenSearch.Client;
+using OpenSearch.Client.Specification.NodesApi;
 using Tests.Core.ManagedOpenSearch;
 using Tests.Core.ManagedOpenSearch.Clusters;
 using Tests.Domain;
