+++ tests/Tests/Cluster/NodesStats/NodesStatsUrlTests.cs
@@ -30,6 +30,7 @@ using System.Threading.Tasks;
 using OpenSearch.OpenSearch.Xunit.XunitPlumbing;
 using OpenSearch.Net;
 using OpenSearch.Client;
+using OpenSearch.Client.Specification.NodesApi;
 using Tests.Framework.EndpointTests;
 using static Tests.Framework.EndpointTests.UrlTester;
